# docs/接口文档/微友App-openapi.yaml
openapi: 3.0.3
info:
  title: 微友 App Backend API Draft
  version: 0.1.0-draft
  description: |
    基于当前后端控制器、需求文档与已生成接口文档整理的 OpenAPI 3.0 草稿。
    
    说明：
    - HTTP Base URL: `http://localhost:8080/api`
    - WebSocket URL: `ws://localhost:8081/ws/chat`
    - WebSocket 协议详情见 `docs/接口文档/微友App-后端接口文档.md`
    - 当前部分接口仍为 demo/stub/no-op 实现，字段结构可用于联调，业务结果未必代表最终生产行为
servers:
  - url: http://localhost:8080/api
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /user/profile/detail:
    get:
      tags:
        - User
      summary: 查询用户资料详情
      description: userId 为空时查询当前登录用户。
      operationId: userProfileDetail
      parameters:
        - in: query
          name: userId
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserProfileData'
  /user/profile/update:
    post:
      tags:
        - User
      summary: 更新资料
      operationId: userProfileUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /user/qrcode/get:
    get:
      tags:
        - User
      summary: 获取个人二维码
      operationId: userQrcodeGet
      parameters:
        - in: query
          name: dynamic
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseQrcodeData'
  /user/status/update:
    post:
      tags:
        - User
      summary: 更新状态
      description: expireAt 当前未实际落库。
      operationId: userStatusUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
  /search/suggest:
    get:
      tags:
        - Search
      summary: 搜索联想词
      description: 当前返回 demo 联想结果。
      operationId: searchSuggest
      parameters:
        - in: query
          name: keyword
          required: true
          schema:
            type: string
        - in: query
          name: bizType
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseStringList'
  /search/global:
    get:
      tags:
        - Search
      summary: 全局搜索
      description: 当前返回 demo 搜索结果。
      operationId: searchGlobal
      parameters:
        - in: query
          name: keyword
          required: true
          schema:
            type: string
        - in: query
          name: bizType
          required: false
          schema:
            type: string
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            format: int32
            default: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageSearchItemData'
...
  /official/account/list:
    get:
      tags:
        - Official
      summary: 查询公众号列表
      operationId: officialAccountList
      parameters:
        - in: query
          name: keyword
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOfficialAccountDetailList'
  /official/account/detail:
    get:
      tags:
        - Official
      summary: 查询公众号详情
      description: 当前返回 demo 数据。
      operationId: officialAccountDetail
      parameters:
        - in: query
          name: officialId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOfficialAccountDetail'
  /official/article/history:
    get:
      tags:
        - Official
      summary: 查询公众号历史文章
      operationId: officialArticleHistory
      parameters:
        - in: query
          name: officialId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: cursor
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageArticleItemData'
  /official/article/detail:
    get:
      tags:
        - Official
      summary: 查询公众号文章详情
      operationId: officialArticleDetail
      parameters:
        - in: query
          name: articleId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseArticleDetail'
  /official/article/like:
    post:
      tags:
        - Official
      summary: 点赞公众号文章
      operationId: officialArticleLike
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleLikeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseArticleLikeResult'
  /official/account/follow:
    post:
      tags:
        - Official
      summary: 关注或取消关注公众号
      operationId: officialAccountFollow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfficialFollowRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOfficialFollowResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponseBase:
      type: object
      required:
        - code
        - message
        - traceId
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        traceId:
          type: string
          example: 3f4b9d6e8a2c4f1b
...
    PageDataBase:
      type: object
      required:
        - pageNo
        - pageSize
        - total
        - hasMore
      properties:
        pageNo:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
        hasMore:
          type: boolean
        nextCursor:
          type: string
          nullable: true
...
    ApiResponseVoid:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              nullable: true
...
    UserProfileData:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        weiyouNo:
          type: string
        nickname:
          type: string
        avatar:
          type: string
        gender:
          type: integer
          format: int32
        city:
          type: string
        signature:
          type: string
        statusText:
          type: string
        qrcodeUrl:
          type: string
    UpdateProfileRequest:
      type: object
      properties:
        nickname:
          type: string
        avatar:
          type: string
        gender:
          type: integer
          format: int32
        city:
          type: string
        signature:
          type: string
    QrcodeData:
      type: object
      properties:
        ticket:
          type: string
        qrcodeUrl:
          type: string
        dynamic:
          type: boolean
        expireAt:
          type: string
    UpdateStatusRequest:
      type: object
      required:
        - statusCode
      properties:
        statusCode:
          type: string
        statusText:
          type: string
        expireAt:
          type: string
...
    SearchItem:
      type: object
      properties:
        bizType:
          type: string
        bizId:
          type: string
        title:
          type: string
        subtitle:
          type: string
        cover:
          type: string
        routePath:
          type: string
...
    OfficialAccountDetail:
      type: object
      properties:
        officialId:
          type: integer
          format: int64
        name:
          type: string
        avatar:
          type: string
        intro:
          type: string
        verified:
          type: boolean
        followed:
          type: boolean
    ArticleItem:
      type: object
      properties:
        articleId:
          type: integer
          format: int64
        officialId:
          type: integer
          format: int64
        title:
          type: string
        summary:
          type: string
        cover:
          type: string
        publishAt:
          type: string
        likeCount:
          type: integer
          format: int32
    ArticleDetail:
      type: object
      properties:
        articleId:
          type: integer
          format: int64
        officialId:
          type: integer
          format: int64
        title:
          type: string
        summary:
          type: string
        contentHtml:
          type: string
        cover:
          type: string
        publishAt:
          type: string
        likeCount:
          type: integer
          format: int32
    ArticleLikeRequest:
      type: object
      required:
        - articleId
        - action
      properties:
        articleId:
          type: integer
          format: int64
        action:
          type: string
    ArticleLikeResult:
      type: object
      properties:
        articleId:
          type: integer
          format: int64
        action:
          type: string
        likeCount:
          type: integer
          format: int32
    OfficialFollowRequest:
      type: object
      required:
        - officialId
        - action
      properties:
        officialId:
          type: integer
          format: int64
        action:
          type: string
    OfficialFollowResult:
      type: object
      properties:
        officialId:
          type: integer
          format: int64
        action:
          type: string
        followed:
          type: boolean
...
    PageSearchItemData:
      allOf:
        - $ref: '#/components/schemas/PageDataBase'
        - type: object
          properties:
            list:
              type: array
              items:
                $ref: '#/components/schemas/SearchItem'
...
    PageArticleItemData:
      allOf:
        - $ref: '#/components/schemas/PageDataBase'
        - type: object
          properties:
            list:
              type: array
              items:
                $ref: '#/components/schemas/ArticleItem'
...
    ApiResponseUserProfileData:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserProfileData'
    ApiResponseQrcodeData:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/QrcodeData'
...
    ApiResponseStringList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                type: string
    ApiResponsePageSearchItemData:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageSearchItemData'
...
    ApiResponseOfficialAccountDetailList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/OfficialAccountDetail'
    ApiResponseOfficialAccountDetail:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OfficialAccountDetail'
    ApiResponsePageArticleItemData:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageArticleItemData'
...
    ApiResponseArticleDetail:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ArticleDetail'
    ApiResponseArticleLikeResult:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ArticleLikeResult'
    ApiResponseOfficialFollowResult:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OfficialFollowResult'
...